LABS/CS10-flight-log/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
django-rest-multiple-models==2.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.6
Pillow==5.2.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.3
requests==2.18.4
sqlparse==0.2.4
stripe==2.10.0
urllib3==1.22
whitenoise==4.1